Refactoring code UI code
* Started to move promptRegister to the helper.
* Made it so you can set the title and label in login-prompt.
