+++ src/login-prompt/login-prompt.html
@@ -3,7 +3,6 @@
 <html class="text-gray-900 dark:text-white">
   <head>
     <meta charset="UTF-8" />
-    <title>Save your Plex Credentials</title>
     <!-- https://electronjs.org/docs/tutorial/security#csp-meta-tag -->
     <meta
       http-equiv="Content-Security-Policy"
@@ -18,24 +17,21 @@
   <body>
     <section class="bg-gray-50 dark:bg-gray-900">
       <div class="mx-auto flex flex-col items-center justify-center px-6 py-8 md:h-screen lg:py-0">
-        <h1 class="mb-6 flex items-center text-3xl font-bold">
-          <img
+        <h1 id="title" class="mb-6 flex items-center text-3xl font-bold">
+          <!-- <img
             class="mr-2 h-8 w-8"
             src="https://flowbite.s3.amazonaws.com/blocks/marketing-ui/logo.svg"
             alt="logo"
-          />
-          Save your Plex Credentials
+          /> -->
         </h1>
         <div
           class="w-full rounded-lg bg-white shadow dark:border dark:border-gray-700 dark:bg-gray-800 sm:max-w-md md:mt-0 xl:p-0"
         >
           <div class="space-y-4 p-6 sm:p-8 md:space-y-6">
-            <h2 class="text-xl font-bold leading-tight tracking-tight md:text-2xl">
-              Enter your account credentials
-            </h2>
+            <h2 id="label" class="text-xl font-bold leading-tight tracking-tight md:text-2xl"></h2>
             <form id="login-form" class="space-y-4 md:space-y-6">
               <div>
-                <label for="username-field" class="mb-2 block text-sm font-medium">Your username</label>
+                <label for="username-field" class="mb-2 block text-sm font-medium">Username</label>
                 <input
                   type="text"
                   name="username-field"
